+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/MainCarContext.kt
@@ -10,6 +10,7 @@ import com.mapbox.navigation.core.MapboxNavigationProvider
 import com.mapbox.navigation.core.formatter.MapboxDistanceFormatter
 import com.mapbox.navigation.ui.maneuver.api.MapboxManeuverApi
 import com.mapbox.navigation.utils.internal.JobControl
+import com.mapbox.search.SearchEngine
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
@@ -17,7 +18,8 @@ import kotlinx.coroutines.SupervisorJob
 @OptIn(MapboxExperimental::class)
 class MainCarContext(
     val carContext: CarContext,
-    val mapboxCarMap: MapboxCarMap
+    val mapboxCarMap: MapboxCarMap,
+    val searchEngine: SearchEngine,
 ) {
     val carSettingsStorage = CarSettingsStorage(carContext)