+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/search/SearchCarContext.kt
@@ -3,7 +3,6 @@ package com.mapbox.androidauto.car.search
 import com.mapbox.androidauto.MapboxCarApp
 import com.mapbox.androidauto.car.MainCarContext
 import com.mapbox.androidauto.car.preview.CarRouteRequest
-import com.mapbox.search.MapboxSearchSdk
 
 /**
  * Contains the dependencies for the search feature.
@@ -17,7 +16,7 @@ class SearchCarContext(
 
     /** SearchCarContext **/
     val carSearchEngine = CarSearchEngine(
-        MapboxSearchSdk.getSearchEngine(),
+        mainCarContext.searchEngine,
         MapboxCarApp.carAppLocationService().navigationLocationProvider
     )
     val carRouteRequest = CarRouteRequest(