+++ libnavui-androidauto/CHANGELOG.md
@@ -5,6 +5,7 @@ Mapbox welcomes participation and contributions from everyone.
 ## Unreleased
 #### Features
 #### Bug fixes and improvements
+- Allow injection of your own SearchEngine. [#6042](https://github.com/mapbox/mapbox-navigation-android/pull/6042)
 
 ## androidauto-v0.3.0 - Jun 24, 2022
 ### Changelog